+++ NeuralNetwork.NET/Networks/PublicAPIs/INeuralNetwork.cs
@@ -39,11 +39,11 @@ namespace NeuralNetworkNET.Networks.PublicAPIs
         /// <remarks>This methods processes a single input row and outputs a single result</remarks>
         [Pure, NotNull]
         [CollectionAccess(CollectionAccessType.Read)]
-        double[] Forward([NotNull] double[] input);
+        float[] Forward([NotNull] float[] input);
 
         [Pure, NotNull]
         [CollectionAccess(CollectionAccessType.Read)]
-        double[,] Forward([NotNull] double[,] x);
+        float[,] Forward([NotNull] float[,] x);
 
         /// <summary>
         /// Calculates the cost function for the current instance and the input values
@@ -52,7 +52,7 @@ namespace NeuralNetworkNET.Networks.PublicAPIs
         /// <param name="y">The expected result to use to calculate the error</param>
         [Pure]
         [CollectionAccess(CollectionAccessType.Read)]
-        double CalculateCost([NotNull] double[] input, [NotNull] double[] y);
+        float CalculateCost([NotNull] float[] input, [NotNull] float[] y);
 
         /// <summary>
         /// Calculates the cost function for the current instance and the input values
@@ -61,7 +61,7 @@ namespace NeuralNetworkNET.Networks.PublicAPIs
         /// <param name="y">The expected result to use to calculate the error</param>
         [Pure]
         [CollectionAccess(CollectionAccessType.Read)]
-        double CalculateCost([NotNull] double[,] input, [NotNull] double[,] y);
+        float CalculateCost([NotNull] float[,] input, [NotNull] float[,] y);
 
         /// <summary>
         /// Serializes the network as a JSON string
